Move timezone conversions to timeseries cleaning module
